New: Error messages on import commands

+++ public/cantusdata/management/commands/import_concordance_data.py
@@ -1,16 +1,25 @@
+from django.core.management.base import BaseCommand
 from cantusdata.models import Concordance
-from django.core.management.base import BaseCommand, CommandError
 
 
 class Command(BaseCommand):
     args = ""
+    debug = True
 
     def handle(self, *args, **kwargs):
         """
         """
-        file = open("data_dumps/" + str(args[0]))
-        # Nuke the db concordances
-        Concordance.objects.all().delete()
+        if args:
+            file_name = str(args[0])
+        else:
+            return self.stdout.write("Please provide a file name!")
+        try:
+            file = open("data_dumps/" + file_name)
+        except IOError:
+            return self.stdout.write(u"File {0} does not exist!".format(file_name))
+        if self.debug:
+            # Nuke the db concordances
+            Concordance.objects.all().delete()
         # Every line is a new concordance
         for line in file.readlines():
             # This method is pretty hacky, but it seems to work
